+++ Stocks.General/ConstantsConfig/BankExchangeConsts.cs
@@ -2,6 +2,6 @@
 {
     public class BankExchangeConsts
     {
-        public const double LiborOneMonthRatio = 0.10;
+        public const decimal LiborOneMonthRatio = 0.10m;
     }
 }
