+++ StocksProccesing.Relational/Helpers/StocksDataAcessHelpers.cs
@@ -8,10 +8,11 @@ namespace StocksProccesing.Relational.Helpers
 {
     public static class StocksDataAcessHelpers
     {
-        public static double GatherCurrentPriceByCompany(this StocksMarketContext context, string ticker)
+        public static decimal GatherCurrentPriceByCompany(this StocksMarketContext context, string ticker)
         {
             //TODO: MODIFY THIS
-            var latestTradingDayStart = DateTimeOffset.UtcNow.AddDays(-10).GetClosestPreviousStockMarketDateTime();
+            var latestTradingDayStart = DateTimeOffset.UtcNow.AddDays(-10)
+                .GetClosestPreviousStockMarketDateTime();
 
             var results = context.PricesData
                 .Where(e => e.Date >= latestTradingDayStart && e.CompanyTicker == ticker)
